% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatemicoint.R
\name{print.hatemicoint}
\alias{print.hatemicoint}
\title{Print Method for hatemicoint Objects}
\usage{
\method{print}{hatemicoint}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"hatemicoint"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for hatemicoint Objects
}
