% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatemicoint.R
\name{summary.hatemicoint}
\alias{summary.hatemicoint}
\title{Summary Method for hatemicoint Objects}
\usage{
\method{summary}{hatemicoint}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"hatemicoint"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns a summary list with inference at various significance levels.
}
\description{
Summary Method for hatemicoint Objects
}
